+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/ElementRequired.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { useTranslation } from "@i18n/client";
 
 import { Checkbox } from "app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
-import { FormElementWithIndex } from "@lib/types/form-builder";
+import { FormElementWithIndex } from "@lib/types/form-builder-types";
 
 export const ElementRequired = ({
   item,
