+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/Modal.tsx
@@ -5,7 +5,7 @@ import PropTypes from "prop-types";
 
 import { Button } from "@clientComponents/globals";
 import { Close } from "@serverComponents/icons";
-import { CDSHTMLDialogElement } from "@lib/types/form-builder";
+import { CDSHTMLDialogElement } from "@lib/types/form-builder-types";
 import { useModalStore } from "@lib/store";
 
 interface IModalContext {
