+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/PanelActions.tsx
@@ -5,7 +5,7 @@ import { useMediaQuery } from "usehooks-ts";
 
 import { FormElementTypes } from "@lib/types";
 import { AddElementButton } from "./elements/element-dialog/AddElementButton";
-import { ElementOptionsFilter, RenderMoreFunc } from "@lib/types/form-builder";
+import { ElementOptionsFilter, RenderMoreFunc } from "@lib/types/form-builder-types";
 import {
   ChevronDown,
   ChevronUp,
