+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/RichText.tsx
@@ -2,7 +2,7 @@
 import React from "react";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { RichTextEditor } from "./lexical-editor/RichTextEditor";
-import { LocalizedElementProperties } from "@lib/types/form-builder";
+import { LocalizedElementProperties } from "@lib/types/form-builder-types";
 import { getPath } from "@clientComponents/form-builder/getPath";
 import { useTranslation } from "@i18n/client";
