+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/element-dialog/AddElementButton.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from "@i18n/client";
 import { FormElementTypes } from "@lib/types";
 import { Button } from "@clientComponents/globals";
 import { ElementDialog } from "./ElementDialog";
-import { ElementOptionsFilter } from "@lib/types/form-builder";
+import { ElementOptionsFilter } from "@lib/types/form-builder-types";
 import { AddIcon } from "@serverComponents/icons";
 
 export const AddElementButton = ({
