+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/element-dialog/descriptions/Departments.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "@i18n/client";
 import { ExampleWrapper } from "./ExampleWrapper";
 import { Combobox, Description, Label } from "@clientComponents/forms";
 import { managedData } from "@lib/managedData";
-import { Language } from "@lib/types/form-builder";
+import { Language } from "@lib/types/form-builder-types";
 
 export const Departments = () => {
   const { t, i18n } = useTranslation("form-builder");
