+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/lexical-editor/RichTextEditor.tsx
@@ -2,7 +2,7 @@
 import React, { useCallback, useState, useEffect } from "react";
 import { Editor } from "./Editor";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
-import { Language } from "@lib/types/form-builder";
+import { Language } from "@lib/types/form-builder-types";
 import debounce from "lodash.debounce";
 import { useTranslation } from "@i18n/client";
