+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/question/QuestionDescription.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useTemplateStore } from "@lib/store";
-import { FormElementWithIndex, LocalizedElementProperties } from "@lib/types/form-builder";
+import { FormElementWithIndex, LocalizedElementProperties } from "@lib/types/form-builder-types";
 import React from "react";
 
 export const QuestionDescription = ({
