+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/sub-elements/SubElementModal.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "@i18n/client";
 
 import { ElementProperties } from "@lib/types";
 import { ModalForm, ModalButton } from "../../index";
-import { FormElementWithIndex } from "@lib/types/form-builder";
+import { FormElementWithIndex } from "@lib/types/form-builder-types";
 import { useModalStore, useTemplateStore } from "@lib/store";
 import { Button } from "@clientComponents/globals";
 import { Modal } from "../../Modal";
