+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components/LanguageLabel.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Language } from "@lib/types/form-builder";
+import { Language } from "@lib/types/form-builder-types";
 import React, { ReactElement } from "react";
 
 // Original: absolute right-0 text-sm top-0 pl-2 pr-2 bg-purple-100 border-purple-200 border mr-1 -mt-6
