+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components/RichText.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "@i18n/client";
 import React from "react";
 import { RichTextEditor } from "../../components/elements/lexical-editor/RichTextEditor";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
-import { Language, LocalizedElementProperties } from "@lib/types/form-builder";
+import { Language, LocalizedElementProperties } from "@lib/types/form-builder-types";
 import { LanguageLabel } from "./LanguageLabel";
 import { FieldsetLegend } from "./FieldsetLegend";
