+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/DownloadSingleButton.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { useTranslation } from "@i18n/client";
 import { getSubmissionsByFormat } from "../actions";
 import { DownloadFormat, HtmlResponse } from "@lib/responseDownloadFormats/types";
-import { Language } from "@lib/types/form-builder";
+import { Language } from "@lib/types/form-builder-types";
 import { usePathname } from "next/navigation";
 
 export const DownloadSingleButton = ({
