+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/SetResponseDelivery.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { useCallback, useState, useMemo } from "react";
-import { LocalizedFormProperties } from "@lib/types/form-builder";
+import { LocalizedFormProperties } from "@lib/types/form-builder-types";
 import axios from "axios";
 import { useTranslation } from "@i18n/client";
 import { useSession } from "next-auth/react";
