+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/navigation/FileName.tsx
@@ -5,7 +5,7 @@ import { cn } from "@lib/utils";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { useTranslation } from "@i18n/client";
 import { useRefStore } from "@lib/hooks/useRefStore";
-import { LocalizedFormProperties } from "@lib/types/form-builder";
+import { LocalizedFormProperties } from "@lib/types/form-builder-types";
 
 export const FileNameInput = () => {
   const { t } = useTranslation(["form-builder"]);
