+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/Dialog.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { useCallback, useEffect, useRef, useState } from "react";
-import { CDSHTMLDialogElement } from "@lib/types/form-builder";
+import { CDSHTMLDialogElement } from "@lib/types/form-builder-types";
 import { useTranslation } from "@i18n/client";
 import { Button } from "@clientComponents/globals";
 import { Close } from "@serverComponents/icons/Close";
