+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/conditionals/AddOther.tsx
@@ -3,7 +3,7 @@ import React, { useCallback } from "react";
 import { useTranslation } from "@i18n/client";
 
 import { Button } from "@clientComponents/globals";
-import { FormElementWithIndex } from "@lib/types/form-builder";
+import { FormElementWithIndex } from "@lib/types/form-builder-types";
 import { useTemplateStore } from "@lib/store";
 import { FormElementTypes } from "@lib/types";
