+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/conditionals/ConditionalIndicator.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React from "react";
 import { useTranslation } from "@i18n/client";
-import { FormElementWithIndex } from "@lib/types/form-builder";
+import { FormElementWithIndex } from "@lib/types/form-builder-types";
 import { useTemplateStore } from "@lib/store";
 import { ConditionalIcon } from "@serverComponents/icons/ConditionalIcon";
 import { getQuestionNumber } from "@formbuilder/util";
