+++ components/clientComponents/globals/Brand.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { BrandProperties } from "@lib/types/form-types";
 import { getLocalizedProperty } from "@lib/utils";
 import { useTranslation } from "@i18n/client";
-import { Language } from "@lib/types/form-builder";
+import { Language } from "@lib/types/form-builder-types";
 
 const Brand = ({
   brand,
