+++ lib/hooks/useAllowPublish.tsx
@@ -3,7 +3,7 @@ import { useCallback, useMemo } from "react";
 
 import { FormElement, FormElementTypes, FormProperties, PropertyChoices } from "@lib/types";
 import { useAccessControl } from "@lib/hooks";
-import { Description, publishRequiredFields, Title } from "../types/form-builder";
+import { Description, publishRequiredFields, Title } from "../types/form-builder-types";
 import { useTemplateStore } from "../store/useTemplateStore";
 
 export class MissingTranslation extends Error {}
